// web/src/features/oauth2/data/schema.ts
//
// Copyright (c) 2025 rustmailer.com (https://rustmailer.com)
//
// This file is part of the Bichon Email Archiving Project
//
// This program is free software: you can redistribute it and/or modify
// it under the terms of the GNU Affero General Public License as published by
// the Free Software Foundation, either version 3 of the License, or
// (at your option) any later version.
//
// This program is distributed in the hope that it will be useful,
// but WITHOUT ANY WARRANTY; without even the implied warranty of
// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
// GNU Affero General Public License for more details.
//
// You should have received a copy of the GNU Affero General Public License
// along with this program.  If not, see <http://www.gnu.org/licenses/>.


export interface OAuth2Entity {
  id: number;
  description?: string;
  client_id: string;
  client_secret: string;
  auth_url: string;
  token_url: string;
  redirect_uri: string;
  scopes?: string[];
  extra_params?: Record<string, string>;
  enabled: boolean;
  use_proxy?: number;
  created_at: number;
  updated_at: number;
}